float to int

+++ example/echo/my_echo.go
@@ -248,10 +248,10 @@ func myPrint(latency_series string) {
 	fmt.Print(latency_series)
 }
 
-func writeThroughput(throughput int64, rate int) {
+func writeThroughput(throughput int64, rate float64) {
 	f, err := os.OpenFile(filename, os.O_APPEND|os.O_WRONLY, 0600)
 	if err != nil {
-		f, err := os.Create(filename + strconv.Itoa(rate) + "/throughputs.txt")
+		f, err := os.Create(filename + strconv.Itoa(int(rate)) + "/throughputs.txt")
 		if err != nil {
 			panic(err)
 		}
